refactor(parser-v3): Move parser utility methods to separate file
- parseFunctionDeclaration
- parseVariableDeclaration
- parseAndMergeKeywords
- updateType
- Remove unused methods:
- parseObjectProperty
