+++ lib/v3/parser.js
@@ -13,6 +13,13 @@ const {
     getAstDefaultOptions
 } = require('../options');
 
+const {
+    parseFunctionDeclaration,
+    parseVariableDeclaration,
+    parseAndMergeKeywords,
+    updateType
+} = require('./v3-utils');
+
 const hasOwnProperty = utils.hasOwnProperty;
 
 const SUPPORTED_FEATURES = [
@@ -30,16 +37,6 @@ const SUPPORTED_FEATURES = [
 const SCOPE_DEFAULT = 'default';
 const SCOPE_STATIC = 'static';
 const SCOPE_MARKUP = 'markup';
-
-const PARAM_ALIASES = {
-    arg: true,
-    argument: true,
-    param: true
-};
-const RETURN_ALIASES = {
-    return: true,
-    returns: true,
-};
 
 class Parser extends EventEmitter {
     constructor(options) {
@@ -123,18 +120,6 @@ class Parser extends EventEmitter {
         }
     }
 
-    updateType(item) {
-        const typeKeyword = item.keywords.find(kw => kw.name === 'type');
-
-        if (typeKeyword) {
-            const parsedType = jsdoc.parseTypeKeyword(typeKeyword.description);
-
-            if (parsedType) {
-                item.type = parsedType;
-            }
-        }
-    }
-
     emitDataItem(variable, parseContext, defaultVisibility, parentComment) {
         const comment = parentComment || utils.getCommentFromSourceCode(variable.node, parseContext.sourceCode, { defaultVisibility });
 
@@ -160,7 +145,7 @@ class Parser extends EventEmitter {
             }];
         }
 
-        this.updateType(item);
+        updateType(item);
 
         this.emit('data', item);
     }
@@ -168,7 +153,7 @@ class Parser extends EventEmitter {
     emitMethodItem(method, parseContext, defaultVisibility, parentComment) {
         const comment = parentComment || utils.getCommentFromSourceCode(method.node, parseContext.sourceCode, { defaultVisibility });
 
-        this.parseKeywords(comment.keywords, method);
+        parseAndMergeKeywords(comment.keywords, method);
 
         const item = Object.assign({}, comment, {
             name: method.name,
@@ -201,7 +186,7 @@ class Parser extends EventEmitter {
             }];
         }
 
-        this.updateType(item);
+        updateType(item);
 
         this.emit('computed', item);
     }
@@ -318,7 +303,7 @@ class Parser extends EventEmitter {
             }
 
             if (node.type === 'VariableDeclaration' && parseContext.scopeType !== SCOPE_MARKUP) {
-                const variables = this.parseVariableDeclaration(node);
+                const variables = parseVariableDeclaration(node);
 
                 variables.forEach(variable => {
                     if (level === 0) {
@@ -356,7 +341,7 @@ class Parser extends EventEmitter {
             }
 
             if (node.type === 'FunctionDeclaration') {
-                this.emitMethodItem(this.parseFunctionDeclaration(node), parseContext, 'private');
+                this.emitMethodItem(parseFunctionDeclaration(node), parseContext, 'private');
 
                 if (node.body) {
                     this.parseBodyRecursively(node.body, parseContext, level + 1);
@@ -373,7 +358,7 @@ class Parser extends EventEmitter {
                     const exportNodeComment = utils.getCommentFromSourceCode(node, parseContext.sourceCode, { defaultVisibility: 'public', useLeading: true, useTrailing: false });
 
                     if (declaration.type === 'VariableDeclaration') {
-                        const variables = this.parseVariableDeclaration(declaration);
+                        const variables = parseVariableDeclaration(declaration);
 
                         variables.forEach(variable => {
                             this.emitDataItem(variable, parseContext, 'public', exportNodeComment);
@@ -403,7 +388,7 @@ class Parser extends EventEmitter {
                     }
 
                     if (declaration.type === 'FunctionDeclaration') {
-                        const func = this.parseFunctionDeclaration(declaration);
+                        const func = parseFunctionDeclaration(declaration);
 
                         this.emitMethodItem(func, parseContext, 'public', exportNodeComment);
 
@@ -581,7 +566,8 @@ class Parser extends EventEmitter {
         const regex = /^{?\s*function\s+\(/i;
 
         expression = expression.replace(regex, function (m) {
-            // When quotes in attributes used curcly braces are provided here, so we should handle it separatly
+            // When quotes in attributes used curly braces are provided here,
+            // so we should handle it separately
             if (m.startsWith('{')) {
                 return '{function a(';
             }
@@ -644,92 +630,6 @@ class Parser extends EventEmitter {
         };
     }
 
-    parseVariableDeclaration(node) {
-        if (node.type !== 'VariableDeclaration') {
-            throw new Error('Node should have a VariableDeclarationType, but is ' + node.type);
-        }
-
-        const result = [];
-
-        node.declarations.forEach(declarator => {
-            const idNode = declarator.id;
-
-            if (idNode.type === 'Identifier') {
-                result.push({
-                    name: idNode.name,
-                    kind: node.kind,
-                    node: node,
-                    declarator: declarator,
-                    location: {
-                        start: idNode.start,
-                        end: idNode.end
-                    }
-                });
-            } else if (idNode.type === 'ObjectPattern') {
-                idNode.properties.forEach(propertyNode => {
-                    const propertyIdNode = propertyNode.key;
-
-                    if (propertyIdNode.type === 'Identifier') {
-                        result.push({
-                            name: propertyIdNode.name,
-                            kind: node.kind,
-                            node: node,
-                            declarator: declarator,
-                            locations: {
-                                start: propertyIdNode.start,
-                                end: propertyIdNode.end
-                            }
-                        });
-                    }
-                });
-            }
-        });
-
-        return result;
-    }
-
-    parseFunctionDeclaration(node) {
-        if (node.type !== 'FunctionDeclaration') {
-            throw new Error('Node should have a FunctionDeclarationType, but is ' + node.type);
-        }
-
-        const params = [];
-
-        node.params.forEach((param) => {
-            if (param.type === 'Identifier') {
-                params.push({
-                    name: param.name,
-                });
-            }
-        });
-
-        const output = {
-            node: node,
-            name: node.id.name,
-            location: {
-                start: node.id.start,
-                end: node.id.end
-            },
-            params: params,
-        };
-
-        return output;
-    }
-
-    parseObjectProperty(node) {
-        if (node.type !== 'Property') {
-            throw new Error('Node should have a Property, but is ' + node.type);
-        }
-
-        if (node.key && node.key.type !== 'Identifier') {
-            throw new Error('Wrong property declaration');
-        }
-
-        return {
-            name: node.key.name
-        };
-    }
-
     parseTemplate() {
         let rootElementIndex = 0;
         let lastComment = null;
@@ -758,7 +658,7 @@ class Parser extends EventEmitter {
 
                 if (this.features.includes('events')) {
                     if (lastTagName !== 'slot') {
-                        // Expose events that propogated from child events
+                        // Expose events that propagated from child events
                         // Handle event syntax like ```<button on:click>Some link</button>```
                         if (name.length > 3 && name.indexOf('on:') === 0 && !value) {
                             const nameWithModificators = name.substr(3).split('|');
@@ -785,7 +685,7 @@ class Parser extends EventEmitter {
                             if (!hasOwnProperty(this.eventsEmitted, baseEvent.name)) {
                                 this.eventsEmitted[baseEvent.name] = baseEvent;
 
-                                this.parseKeywords(comment.keywords, baseEvent);
+                                parseAndMergeKeywords(comment.keywords, baseEvent);
 
                                 this.emit('event', baseEvent);
                             }
@@ -915,49 +815,6 @@ class Parser extends EventEmitter {
 
         parser.write(this.structure.template);
         parser.end();
-    }
-
-    /**
-     * Mutates event.
-     * @param {any[]} keywords
-     * @param {{ params?: any[] }} event
-     */
-    parseKeywords(keywords = [], event) {
-        if (!event.params) {
-            event.params = [];
-        }
-
-        keywords.forEach(({ name, description }) => {
-            if (name in PARAM_ALIASES) {
-                const parsedParam = jsdoc.parseParamKeyword(description);
-                const pIndex = event.params.findIndex(
-                    p => p.name === parsedParam.name
-                );
-
-                /*
-                 * Replace the param if there is already one present with
-                 * the same name. This will happen with parsed
-                 * FunctionDeclaration because params will already be
-                 * present from parsing the AST node.
-                 */
-                if (pIndex >= 0) {
-                    event.params[pIndex] = parsedParam;
-                } else {
-                    /*
-                     * This means @param does not match an actual param
-                     * in the FunctionDeclaration.
-                     * TODO: Implement option to choose behaviour (keep, ignore, warn, throw)
-                     */
-                    event.params.push(parsedParam);
-                }
-            } else if (name in RETURN_ALIASES) {
-                event.return = jsdoc.parseReturnKeyword(description);
-            }
-        });
-
-        if (event.params.length === 0) {
-            delete event.params;
-        }
     }
 }
